# lib/api-zod/src/generated/types/briefingSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API pour l'outil d'analyse météorologique bihedomadaire NOAA
 * OpenAPI spec version: 0.1.0
 */
import type { BriefingContentType } from './briefingContentType';
import type { BriefingSubsection } from './briefingSubsection';
import type { IframeModel } from './iframeModel';

export interface BriefingSection {
  id: string;
  name: string;
  contentType: BriefingContentType;
  subsections: BriefingSubsection[];
  iframeUrl?: string;
  iframeAlternatives?: IframeModel[];
  externalUrl?: string;
  externalLabel?: string;
  externalDescription?: string;
}