# lib/api-zod/src/generated/types/chartSubsection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API pour l'outil d'analyse météorologique bihedomadaire NOAA
 * OpenAPI spec version: 0.1.0
 */
import type { ChartItem } from './chartItem';

export interface ChartSubsection {
  label: string;
  charts: ChartItem[];
}