# lib/api-zod/src/generated/types/noaaSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API pour l'outil d'analyse météorologique bihedomadaire NOAA
 * OpenAPI spec version: 0.1.0
 */
import type { ChartSubsection } from './chartSubsection';

export interface NoaaSection {
  id: string;
  name: string;
  subsections: ChartSubsection[];
}